+++ examples/customizations/marytts_mp3_generator/__init__.py
@@ -4,7 +4,7 @@ import tempfile
 import ffmpeg
 from speak2mary import MaryTTS
 
-from faker_file.providers.mp3_file.generators.base import BaseMp3Generator
+from faker_file.providers.base.mp3_generator import BaseMp3Generator
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2022-2023 Artur Barseghyan"