+++ src/faker_file/providers/mp3_file/generators/gtts_generator.py
@@ -2,7 +2,7 @@ from io import BytesIO
 
 from gtts import gTTS
 
-from .base import BaseMp3Generator
+from ...base.mp3_generator import BaseMp3Generator
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2022-2023 Artur Barseghyan"